make maintenance scripts start even when not connected
to run `help <task>` the shell wrapper `maintenace.sh` will invoke the
arangosh with the extra parameter `--server.endpoint none`. when there
is no endpoint, calling `arango.isConnected()` will unconditionally
throw an exception, so that the startup of the maintenance scripts was
aborted.
Using a try...catch here makes the startup go on, so that `help <task>`
works again.